+++ apps/e2e-tests/page-objects/WishlistPage.ts
@@ -5,6 +5,7 @@ export class WishlistPage {
   readonly addToCartButton: Locator;
   readonly wishlistButton: Locator;
   readonly productInWishlistButton: Locator;
+  readonly clearWishlistButton: Locator;
 
   constructor(page: Page) {
     this.page = page;
@@ -13,6 +14,7 @@ export class WishlistPage {
     this.productInWishlistButton = page
       .getByTestId("product-box-toggle-wishlist-button")
       .first();
+    this.clearWishlistButton = page.getByTestId("clear-wishlist-button");
   }
 
   async openWishlist() {
@@ -22,5 +24,10 @@ export class WishlistPage {
   async removeProductFromWishlist() {
     await this.page.waitForLoadState("networkidle");
     await this.productInWishlistButton.click();
+  }
+
+  async clearWishlist() {
+    await this.page.waitForLoadState("networkidle");
+    await this.clearWishlistButton.click();
   }
 }