+++ packages/api-client-next/admin-api-types/index.d.ts
@@ -1,0 +1,1 @@
+export * from "./apiTypes-6.5.3.0";