+++ packages/api-client-next/build.config.ts
@@ -2,5 +2,9 @@ import { defineBuildConfig } from "unbuild";
 export default defineBuildConfig({
   entries: ["src/index"],
   declaration: true,
+  rollup: {
+    cjsBridge: true,
+    emitCJS: true,
+  },
   externals: [],
 });