+++ packages/composables/src/cms/useCmsElementImage.ts
@@ -12,12 +12,19 @@ import {
   ImgHTMLAttributes,
 } from "vue";
 import { getSrcSetForMedia } from "@shopware-pwa/helpers-next";
+import { urlIsAbsolute } from "../helpers/urlIsAbsolute";
+
+export type ImageContainerAttrs = {
+  href?: string;
+  target?: string;
+  rel?: string;
+};
 
 export type UseCmsElementImage = {
   containerStyle: ComputedRef<CSSProperties>;
   anchorAttrs: ComputedRef<AnchorHTMLAttributes>;
   imageAttrs: ComputedRef<ImgHTMLAttributes>;
-  imageContainerAttrs: ComputedRef<CSSProperties>;
+  imageContainerAttrs: ComputedRef<ImageContainerAttrs>;
   imageLink: ComputedRef<{ newTab: boolean; url: string }>;
   displayMode: ComputedRef<DisplayMode>;
 };
@@ -47,9 +54,11 @@ export function useCmsElementImage(
   }));
 
   const imageContainerAttrs = computed(() => {
-    const attr: { [k: string]: string } = {};
+    const attr: ImageContainerAttrs = {};
     if (imageLink.value.url) {
-      attr.href = imageLink.value.url;
+      attr.href = urlIsAbsolute(imageLink.value.url)
+        ? imageLink.value.url
+        : `/${imageLink.value.url}`;
     }
     if (imageLink.value.newTab) {
       attr.target = "blank";