+++ templates/vue-demo-store/pages/wishlist.vue
@@ -71,6 +71,7 @@ watch(
 </script>
 
 <template>
+  <LayoutBreadcrumbs />
   <div class="wishlist-page" data-testid="wishlist-wrapper">
     <div
       class="max-w-2xl mx-auto py-4 px-4 sm:py-4 sm:px-6 lg:max-w-7xl lg:px-8"